Synchronize players
I deviated from the course as I thought his code was too complicated. The principle of setting important values like global position only on the host machine, though, is very sound.

What I did here was to make each player their own multiplayer authority, gather input in _process() only on the authoritative machine, and then apply movement to all peers in _physics_process() only on the host machine.

The set process and physics process calls could easily be replaced with checks in the respective functions if more logic is added to those callbacks that requires crossing the boundaries a bit.