Rework bullet synchronization slightly
This is a bit confusing but even though I went off-course with my architecture, the course helped out a ton here.

The bullets are being spawned on every client by the host's MultiplayerSpawner. That means their _physics_process() runs for every client, which means global_position (which is updated in there) does not need to be synchronized after spawn.

The rotation and movement direction also don't have to be synchronized after spawn. However, the movement direction does need to be synchronized so newly spawned bullets move in the right direction on each client.